refactor(portfolio): address SonarQube findings in ProjectModal + card
Split ProjectModal into three files to match ExperienceModal's
Modal/ModalHeader/ModalContent decomposition, which also drops the
cognitive complexity from 23 below SonarQube's 15 threshold (S3776):

- ProjectModal.tsx        (129 lines, shell + portal only)
- ProjectModalHeader.tsx  (new, sticky header)
- ProjectModalBody.tsx    (new, description/stack/features/contributors/links)

Extract two helpers in portfolioConstants.ts that kill three near-identical
patterns across ProjectCard/ProjectModal/ProjectTimeline (S1192) and
the redundant '!== ""' empty-string checks (S1764):

- getCategoryColors(category)  -- replaces three copies of
  CATEGORY_COLORS[x] || CATEGORY_COLORS.Others and switches || to ??
- isValidUrl(url)              -- replaces hasGithub/hasLive guards
